refactor: LottoTicket 제거 및 기능 Lotto 클래스로 이동
