+++ src/main/java/lotto/domain/LottoTickets.java
@@ -4,13 +4,13 @@ import java.util.List;
 
 public class LottoTickets {
 
-    private final List<LottoTicket> lottoTickets;
+    private final List<Lotto> lottos;
 
-    public LottoTickets(List<LottoTicket> lottoTickets) {
-        this.lottoTickets = lottoTickets;
+    public LottoTickets(List<Lotto> lottos) {
+        this.lottos = lottos;
     }
 
-    public List<LottoTicket> getLottoTickets() {
-        return lottoTickets;
+    public List<Lotto> getLottos() {
+        return lottos;
     }
 }
